+++ core/CHANGELOG.md
@@ -27,12 +27,13 @@
 - 🆕 [v1.0.0] `components/functionalities/theFallenOneFunctionalityBuilder`.
 - 🆕 [v1.0.0] `engineGenerator`.
 - 🆕 [v1.0.0] `engineGenerator/beforeStopExecutionCheckers/starvationCheckerBuilder`.
-- 🆕 [v1.0.0] `engineGenerator/processors/emotionResetProcessorBuilder`.
-- 🆕 [v1.0.0] `engineGenerator/processors/positioningProcessorBuilder`.
-- 🆕 [v1.0.0] `engineGenerator/processors/processorGenerator`.
 - 🆕 [v1.0.0] `engineGenerator/processors/checkers/engineErrorBuilder`.
 - 🆕 [v1.0.0] `engineGenerator/processors/checkers/hasBecomeTheFallenOneBuilder`.
 - 🆕 [v1.0.0] `engineGenerator/processors/checkers/hasHitAWallBuilder`.
+- 🆕 [v1.0.0] `engineGenerator/processors/emotionResetProcessorBuilder`.
+- 🆕 [v1.0.0] `engineGenerator/processors/positioningProcessorBuilder`.
+- 🆕 [v1.0.0] `engineGenerator/processors/processorGenerator`.
+- 🆕 [v1.0.0] `engineGenerator/processors/theFallenOneResetProcessorBuilder`.
 - 🆕 [v1.0.0] `generics/actorGenerator`.
 
 ### 💅 React UI